requirements.txt
bs4==0.0.1
imdbpy==2021.4.18
ipython==5.5.0
kaleido==0.2.1
Summary: Stati
numpy==1.19.5
pandas==1.1.5
plotly==5.1.0
python==3.7.11
regex==2019.12.20
requests==2.23.0
requests-oauthlib==1.3.0
tqdm==4.41.1
